[axios] Add error handling before all axios uses

+++ src/components/buttons/users/DeleteAccount.js
@@ -13,14 +13,16 @@ function DeleteAccount(props) {
     }
 
     const handleConfirmClicked = async (e) => {
-        try {
-            e.preventDefault()
-            await axios.delete(
-                REACT_APP_SERVER_URL + 'users/' + props.user.id
-            )
-            setConfirmClicked(true)
-        } catch(error) {
-            alert(error.response.data.msg)
+        e.preventDefault()
+        if (props.user.id) {
+            try {
+                await axios.delete(
+                    REACT_APP_SERVER_URL + 'users/' + props.user.id
+                )
+                setConfirmClicked(true)
+            } catch(error) {
+                alert(error.response.data.msg)
+            }
         }
     }
